requirements.txt
aiohappyeyeballs @ file:///home/conda/feedstock_root/build_artifacts/aiohappyeyeballs_1741775197943/work
aiohttp @ file:///Users/runner/miniforge3/conda-bld/aiohttp_1745255757319/work
aiosignal @ file:///home/conda/feedstock_root/build_artifacts/aiosignal_1734342155601/work
annotated-types==0.7.0
anyio==4.9.0
async-timeout @ file:///home/conda/feedstock_root/build_artifacts/async-timeout_1733235340728/work
attrs @ file:///home/conda/feedstock_root/build_artifacts/attrs_1741918516150/work
certifi==2025.4.26
compas @ file:///home/conda/feedstock_root/build_artifacts/compas_1745472428631/work
compas_ifc==1.5.0
compas_occ @ file:///home/conda/feedstock_root/build_artifacts/compas_occ_1737393547869/work
contourpy==1.3.0
cycler==0.12.1
distro==1.9.0
exceptiongroup==1.2.2
fonttools==4.58.0
frozenlist @ file:///Users/runner/miniforge3/conda-bld/frozenlist_1737645340620/work
git-filter-repo==2.47.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna @ file:///home/conda/feedstock_root/build_artifacts/idna_1733211830134/work
ifcopenshell==0.7.0.240406
importlib_resources @ file:///home/conda/feedstock_root/build_artifacts/importlib_resources_1736252299705/work
jiter==0.9.0
joblib==1.5.0
jsonschema @ file:///home/conda/feedstock_root/build_artifacts/jsonschema_1733472696581/work
jsonschema-specifications @ file:///tmp/tmpuvkyqc9y/src
kiwisolver==1.4.7
lark==1.2.2
loguru @ file:///Users/runner/miniforge3/conda-bld/loguru_1725349717395/work
matplotlib==3.9.4
msgpack @ file:///Users/runner/miniforge3/conda-bld/msgpack-python_1725975035107/work
multidict @ file:///Users/runner/miniforge3/conda-bld/multidict_1744972470373/work
networkx @ file:///home/conda/feedstock_root/build_artifacts/networkx_1698504735452/work
numpy @ file:///Users/runner/miniforge3/conda-bld/numpy_1707225581655/work/dist/numpy-1.26.4-cp39-cp39-macosx_10_9_x86_64.whl#sha256=92e1261eb25289fae8429db9951fee08c8ac1116bc17494ff0d57f74d115de46
openai==1.77.0
packaging @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_packaging_1745345660/work
pandas==2.2.3
patsy==1.0.1
pillow==11.2.1
pkgutil_resolve_name @ file:///home/conda/feedstock_root/build_artifacts/pkgutil-resolve-name_1733344503739/work
propcache @ file:///Users/runner/miniforge3/conda-bld/propcache_1744524952218/work
pydantic==2.11.4
pydantic_core==2.33.2
pyDOE2==1.3.0
pyparsing @ file:///home/conda/feedstock_root/build_artifacts/pyparsing_1743089729650/work
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML @ file:///Users/runner/miniforge3/conda-bld/pyyaml_1737454726946/work
referencing @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_referencing_1737836872/work
rpds-py @ file:///Users/runner/miniforge3/conda-bld/bld/rattler-build_rpds-py_1743037541/work
scikit-learn==1.6.1
scipy @ file:///Users/runner/miniforge3/conda-bld/scipy-split_1716470285601/work/dist/scipy-1.13.1-cp39-cp39-macosx_10_13_x86_64.whl#sha256=c9dd6de2b24ca5604e6e5444642497657db2f4e882c088c0d710addb7781a300
seaborn==0.13.2
six @ file:///home/conda/feedstock_root/build_artifacts/six_1733380938961/work
sniffio==1.3.1
statsmodels==0.14.4
svgwrite @ file:///home/conda/feedstock_root/build_artifacts/svgwrite_1734380181028/work
threadpoolctl==3.6.0
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_typing_extensions_1744302253/work
tzdata==2025.2
vtk==9.2.6
watchdog @ file:///Users/runner/miniforge3/conda-bld/watchdog_1730492909178/work
wslink @ file:///home/conda/feedstock_root/build_artifacts/wslink_1742768409749/work
yarl @ file:///Users/runner/miniforge3/conda-bld/yarl_1744972513462/work
zipp @ file:///home/conda/feedstock_root/build_artifacts/zipp_1732827521216/work
